tests: Prevent weird type promotion objection in bgp tests
The va_start function cannot take a object that can be type promoted
Looks like a new warning coming in from a new compiler.

+++ tests/bgpd/test_bgp_table.c
@@ -116,8 +116,8 @@ static void check_lookup_result(struct list *list, va_list arglist)
 	assert(prefix_count == listcount(list));
 }
 
-static void do_test(struct bgp_table *table, const char *prefix, uint8_t maxlen,
-		    ...)
+static void do_test(struct bgp_table *table, const char *prefix,
+		    uint32_t maxlen, ...)
 {
 	va_list arglist;
 	struct list *list = list_new();
